requirements.txt
supabase==2.7.4
requests==2.32.3
python-dotenv==1.0.1
tenacity==9.0.0
gTTS==2.5.3
